server/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
sqlalchemy>=1.4.22
pydantic>=1.8.2
numpy>=1.21.2
pytest>=6.2.5
faker>=13.3.1
pydantic-settings>=2.0.3
pydantic[email]
python-multipart
scikit-learn 
pandas 
joblib
deap